refactor: Move PacketSender to `arcade-utils`
