+++ arcade-virtual-entities/src/main/kotlin/net/casual/arcade/virtual/entity/utils/VirtualEntityPacketCollector.kt
@@ -7,8 +7,8 @@ package net.casual.arcade.virtual.entity.utils
 import it.unimi.dsi.fastutil.ints.IntArrayList
 import it.unimi.dsi.fastutil.objects.ObjectArrayList
 import it.unimi.dsi.fastutil.objects.ObjectListIterator
-import net.casual.arcade.observer.packet.PacketSender
 import net.casual.arcade.utils.asClientGamePacket
+import net.casual.arcade.utils.network.PacketSender
 import net.minecraft.network.protocol.BundlerInfo
 import net.minecraft.network.protocol.Packet
 import net.minecraft.network.protocol.game.ClientGamePacketListener
